+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 11.8.1
 -----------------
 - Fix IE not propagating focus event to parent on click in Inputs.php
+- Fix regression since 8.6 add missing Request Widget to Advanced Student Search in Widgets.php
 
 Changes in 11.8
 ---------------