+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 11.8.1
 - Use 'search' type for Search input in Courses.php
 - Rollback JS set ghost div height to 4000px in jquery-fixedmenu.js
 - Fix automatically calculate Final Grades for Progress Periods in FinalGrades.inc.php
+- CSS fix Grade points line height in stylesheet.css
 
 Changes in 11.8
 ---------------