+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes in 11.8.1
 - Rollback JS set ghost div height to 4000px in jquery-fixedmenu.js
 - Fix automatically calculate Final Grades for Progress Periods in FinalGrades.inc.php
 - CSS fix Grade points line height in stylesheet.css
+- Fix SQL error Cannot delete: foreign key constraint fails (marking_period_id + course_periods) in Rollover.php
 
 Changes in 11.8
 ---------------