+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 12.0
+---------------
+- SQL update User password for all school years in PasswordReset.php, Preferences.php & User.php
+
 Changes in 11.8.1
 -----------------
 - Fix IE not propagating focus event to parent on click in Inputs.php